Move ActivityResult into android.events package.
