+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Move `ActivityStore` into `android` package.
 - Move `FragmentId` into `android` package.
 - Move `FormulaAppCompatActivity` into `android` package.
+- Move `ActivityResult` into `android.events` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.
