+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentAndroidEventTest.kt
@@ -4,7 +4,7 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
-import com.instacart.formula.activity.ActivityResult
+import com.instacart.formula.android.events.ActivityResult
 import com.instacart.formula.test.TestFragmentActivity
 import com.instacart.formula.test.TestLifecycleContract
 import io.reactivex.rxjava3.core.Observable
