+++ formula-android/src/main/java/com/instacart/formula/android/ActivityStoreContext.kt
@@ -2,7 +2,7 @@ package com.instacart.formula.android
 
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.Lifecycle
-import com.instacart.formula.activity.ActivityResult
+import com.instacart.formula.android.events.ActivityResult
 import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.fragment.FragmentFlowStore
 import com.instacart.formula.fragment.FragmentKey
